Add indexes to temp tables when updating.
The idnexes which are added are for those columns which
are used to calculate information such as latest upstream version.
These queries use subqueries which seemed to be adversely
impacted by the lack of indexes.

In total, reduced update time on test device from just over 100 seconds
to just over 60 seconds.