use background thread for swap icons instead of AsyncTask
An AsyncTask ties into the UI thread for things like onPostExecute(). If it
is run within an AsyncTask, then it freaks out because it can't tie into
the UI thread.  We don't need it to do that here anyway, so just use a
plain Thread, and set the priority to background.