implement resumable APK downloads for HttpDownloader
Moving towards having the Downloader always download directly into the
cache means its really easy to support resuming downloads.

This should also work for updates and icons, since it fetches the
Content Length via HTTP first.  The icon and update downloading code
needs to be adjusted to support that. For APKs, it probably makes
sense to include the file size from the index db to save the query
over the net.  That would be probably more helpful on Bluetooth.