make db maxSdkValues values use Byte.MAX_VALUE as max not 0
Having 0 mean max makes the logic confusing when maxSdkValue is used in
variable.  This sanitizes the data so that maxSdkValue is always just a
plain int value that can be used to test against.  It does this by
defaulting to Byte.MAX_VALUE (127) if it is not explicitly set. At the rate
of 24 SDK numbers in 8 years, that gives us about 24 years before we have
to think about setting it to Short.MAX_VALUE.

This fixes an issue created by e021eb5ca7e8f05dbce7c1b87833722542138302