add method to check if a URL is being handled by DownloaderService
This also saves the activeDownloadUrlString per packageName. Both are
necessary so that AppDetails can accurately display the current state
of a background download. Saving this per-packageName allows there to
be multiple active downloads in the background, then when people move
around AppDetails, it'll restore the progress meter and button state
when coming back to the app that they clicked install on.

By definition, there is just one DownloaderService enforced by
Android with a single active Downloader instance enforced by the
DownloaderService.  That means using a static variable maps directly
to those conditions and provides a really simple, implementation,
especially compared to what would have to happen to do it via messages
from the thread and any Activities.  If this ends up blocking testing
or something, it can always be changed when someone implements those
tests.