Do not ellipsize description on Android 2.X
For some reason, even when showing the entire description some words at
the end of lines and paragraphs still get cut off. This results in
little parts of the description being completely inaccessible.

Ellipsizing is just cosmetic, so ditch it on 2.X altogether. I tested
and proved that this fixes the issue with an android-10 x86 emulator,
after reproducing the bug.

Keep the ellipsizing on newer versions, as it doesn't misbehave on
those.

Fixes #510.