use Environment.getRootDirectory() instead of /system
Its officially possible to have the ROM's filesystem with any name. While I
have never seen that in practice, Android does provide an easy method to
get the real name.  Plus this should help avoid typos and the like, and
make it easy to track things that rely on that filesystem path.