reorganize swap's App() constructor to prepare for database caching
To generate swap's index.jar, lots of information about all the installed
APKs needs to be parsed.  That can take a long time.  Some of that can be
stored in InstalledAppProvider.  This prepares for those changes.

Also, turns out that packageInfo.applicationInfo provides enough info, so
there is no need to use pm.getApplicationInfo().  And the metadata from
GET_META_DATA was not even being used.