Replace rate limiting code with RX.
Now that we have RX as a dependency, it can be used as a nice concise way to
achieve certain tasks. Rate limiting is one thing it does well - via the
`debounce` mechanism:

  http://reactivex.io/documentation/operators/debounce.html

The semantics of this code is the same as before, limiting content change notifications
to one per second.